+++ service/controller/resource/alerting/alertmanagerservicemonitor/doc.go
@@ -1,0 +1,4 @@
+package alertmanagerservicemonitor
+
+// This packages creates the servicemonitor that makes prometheus
+// able to scrape metrics from alertmanager